requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
onnxruntime>=1.17.0
huggingface-hub>=0.20.0
